requirements.txt
tensorflow==2.4.1 
pandas
numpy
watchdog
json
flask
pillow
